[wfs] Move gui-related stuff to a new file
